# Conflicts:
#	backend/src/main/java/com/recipes/backend/bizz/ingredient/IngredientService.java
#	backend/src/main/java/com/recipes/backend/bizz/ingredient/IngredientServiceImpl.java
#	backend/src/main/java/com/recipes/backend/rest/IngredientController.java
#	backend/src/main/java/com/recipes/backend/rest/LoginController.java
#	backend/src/test/java/com/recipes/backend/bizz/ingredient/IngredientServiceIntegrationTest.java

+++ backend/src/main/java/com/recipes/backend/rest/IngredientController.java
@@ -87,7 +87,9 @@ public class IngredientController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteIngredient(@RequestHeader HttpHeaders headers,
                                                    @PathVariable(name = "id") Long ingredientId) {
-        //TODO add header validation
+        if (!securityService.isAuthenticated(headers)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         return ingredientService.deleteIngredient(ingredientId) ? ResponseEntity.ok(ingredientId.toString()) : ResponseEntity.badRequest().body("Bad request!");
     }
